# openapi-herramientas-equipos.yaml
openapi: 3.0.3
info:
  title: API de Herramientas y Equipos Retirados
  description: >
    Módulo de backend de la app Wifix (Fase 1). Cubre dos dominios del
    servicio de operaciones de campo:

      1. HERRAMIENTAS — almacenamiento de resultados de mediciones:
         medición de distancia, test de velocidad, mapa de calor WiFi,
         ping y traceroute.
      2. EQUIPOS RETIRADOS — registro de equipos retirados o reemplazados.

    CLAVE DE IDENTIFICACIÓN (FASE 1)
    ----------------------------------------------------------------------
    En la Fase 1 la clave práctica de cada registro es `accountNumber`:
    el número de cuenta del cliente que el técnico ingresa en la app.
    Los demás identificadores (`clientId`, `visitId`, `technicianId`) son
    opcionales: provendrán de un sistema upstream aún por definir y en la
    Fase 1 se completan con valores de prueba. `contractId` también es
    opcional y en una fase posterior será reemplazado por `taskId`.

    AUTENTICACIÓN
    ----------------------------------------------------------------------
    La app tiene login propio. POST /auth/login devuelve un JWT que el
    resto de endpoints exige en el header Authorization: Bearer.

    DATOS DE SISTEMAS EXTERNOS (MOCKUP)
    ----------------------------------------------------------------------
    Los campos que provienen de sistemas de la operadora (Comarch/TYTAN,
    FSM, ISP Monitor, ACS, TEC, RMS) se sirven en esta etapa con datos
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /accounts/{accountNumber}/unsatisfactory-tasks:
    get:
      tags: [Tareas y Visitas]
      summary: Observaciones de cierre de tareas insatisfactorias
      description: >
        Tareas insatisfactorias previas con fecha/hora, motivo y observación de
        cierre (campo 15). Origen: FSM `POST /account/process` +
        `POST /workorder/tasks` sobre como mucho `limit` órdenes **finalizadas
        más recientes**. **1 + ≤5 llamadas upstream**, concurrencia 4.


        ⚠ Cambio de forma: antes era un array desnudo, ahora es un objeto con
        `items`, `scanned`, `totalOrders` y `truncated`, para que la UI pueda
        decir "se revisaron las 5 órdenes más recientes de 12" en vez de mentir
        por omisión.


        `result` es un **valor inferido** por `classifyTaskResult()`, no un dato
        de la operadora: la documentación de FSM no enumera los valores de
        `status` ni expone un campo satisfactoria/insatisfactoria.
      operationId: getUnsatisfactoryTasks
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - name: limit
          in: query
          required: false
          description: Órdenes finalizadas más recientes cuyas notas se leen.
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsatisfactoryTasks'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
  /accounts/{accountNumber}/previous-visits:
    get:
      tags: [Tareas y Visitas]
      summary: Motivos y observaciones de visitas anteriores
      description: >
        Visitas anteriores con fecha/hora y motivo (campo 16). Origen: FSM
        `POST /account/process` con `estado:"Todas"`. **1 sola llamada
        upstream, sin notas**: `closingNotes` viene vacío y `notesLoaded` en
        `false`. Las notas se piden por visita con
        `GET /workorders/tasks?workOrder=` al expandirla (ADR-06).


        ⚠ Cambio de forma: antes era un array desnudo, ahora es un objeto con
        `items` y `totalOrders`.
      operationId: getPreviousVisits
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: >-
            OK. Una cuenta sin órdenes devuelve la lista `items` vacía.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviousVisits'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
  /accounts/{accountNumber}/orders:
    get:
      tags: [Tareas y Visitas]
      summary: Órdenes de trabajo de la cuenta
      description: >
        Órdenes crudas normalizadas de FSM `POST /account/process`
        (base de los campos 1-3, 15 y 16; también sirve de diagnóstico).
        **1 llamada upstream.** No devuelve 404: una cuenta sin órdenes es un
        `orders: []` legítimo.
      operationId: getAccountOrders
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - name: estado
          in: query
          required: false
          description: >
            `Todas`: finalizadas y no finalizadas. `Pendientes`: solo las que
            aún no finalizan (`finish_date is null`).
          schema:
            type: string
            enum: [Todas, Pendientes]
            default: Todas
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOrders'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
  /workorders/tasks:
    get:
      tags: [Tareas y Visitas]
      summary: Tareas y notas de una orden de trabajo
      description: >
        Notas de cierre de una orden concreta (campos 15 y 16), desde FSM
        `POST /workorder/tasks`. **1 llamada upstream**, disparada al expandir
        una visita.


        ⚠ `workOrder` va por **query, no por path**: el valor es
        `ORDER/424900/2026` y las barras rompen el ruteo aunque se codifiquen
        como `%2F` (Fastify y varios proxies las normalizan).
      operationId: getWorkOrderTasks
      parameters:
        - name: workOrder
          in: query
          required: true
          description: Número de orden, tal cual lo devuelve `/account/process`.
          schema:
            type: string
            minLength: 1
            maxLength: 120
            example: ORDER/424900/2026
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrderTasks'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'

  # ====================================================================
  # INTEGRACIONES (FSM)
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por POST /auth/login.

  # --------------------------------------------------------------------
  # PARÁMETROS REUTILIZABLES
  # --------------------------------------------------------------------
  parameters:
...
    AccountNumberPath:
      name: accountNumber
      in: path
      required: true
      description: Número de cuenta del cliente.
      schema:
        type: string
...
    BrandHeader:
      name: X-Wifix-Brand
      in: header
      required: false
      description: >
        Marca comercial (realm de FSM) con la que se consulta la operadora.
        Si falta, el backend usa `FSM_DEFAULT_BRAND`. Internamente nunca se
        habla de "realm": los nombres `realm-ecommerce-callcenter-*` quedan
        encerrados en el backend y el técnico no los ve (ADR-01).
      schema:
        type: string
        enum: [telenews, seteinfo]
        example: telenews
    BrandQuery:
      name: brand
      in: query
      required: false
      description: >
        Override de `X-Wifix-Brand` por query. Pensado para el script de sondeo
        y soporte; la webapp usa el header.
      schema:
        type: string
        enum: [telenews, seteinfo]
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Token ausente, inválido o expirado / credenciales incorrectas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConnectorError:
      description: >
        Un sistema externo falló (5xx, timeout o JSON inválido). No es un
        problema de credenciales.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamAuthUnavailable:
      description: >
        El acceso a FSM no está disponible: token ausente, vencido o rechazado.


        **Es 503 a propósito, nunca 401.** La webapp cierra la sesión del
        técnico ante cualquier 401 (`api.js`), y un token caducado de un tercero
        no puede sacarlo al login en medio de una visita ni tumbarle Herramientas
        y Equipos Retirados, que no dependen de FSM.


        El cliente debe pintar un banner no bloqueante en el panel afectado y
        **no** cerrar sesión, no limpiar el token de Wifix y no recargar la app.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpstreamAuthError'

  # --------------------------------------------------------------------
  # ESQUEMAS
  # --------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Código de error legible por máquina.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Descripción del error legible por humanos.
          example: El campo accountNumber es obligatorio.
        details:
          type: array
          description: Detalle por campo, cuando aplica.
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
        meta:
          type: object
          additionalProperties: true
          description: >
            Contexto adicional legible por máquina (integración, marca, motivo).
            Opcional y aditivo: el cliente nunca debe asumir que existe.
    UpstreamAuthError:
      type: object
      description: >
        Error de autenticación contra una API de terceros (hoy FSM). Siempre
        HTTP 503.
      required: [code, message]
      properties:
        code:
          type: string
          enum: [UPSTREAM_AUTH_ERROR]
        message:
          type: string
          description: Texto en español, listo para mostrar al técnico.
          example: >-
            El token de FSM (marca telenews) venció el 05/09/2026 21:59. Pide la
            renovación al contacto de la operadora.
        meta:
          type: object
          properties:
            integration:
              type: string
              example: FSM
            brand:
              type: string
              enum: [telenews, seteinfo]
            reason:
              type: string
              enum: [MISSING, EXPIRED, REJECTED]
              description: >
                `MISSING`: no hay acceso configurado en el servidor.
                `EXPIRED`: el token venció (detectado SIN salir a la red).
                `REJECTED`: la operadora lo rechazó (revocado o de otro realm).
            tokenExpiresAt:
              type: string
              format: date-time
              nullable: true
            retryable:
              type: boolean
              example: false
    Degraded:
      type: object
      description: >
        Aviso de que una respuesta 200 se sirvió con una fuente alternativa o
        incompleta. **No es un error.** Ausente cuando todo salió bien; el
        cliente nunca debe asumir que existe.
      required: [reason, message]
      properties:
        reason:
          type: string
          enum: [FSM_AUTH, FSM_UNAVAILABLE, SOURCE_FALLBACK, TRUNCATED]
        message:
          type: string
          description: Texto en español, listo para mostrar.
...
    ClosedTask:
      type: object
      description: Tarea o visita cerrada, con su observación de cierre.
      properties:
        taskId:
          type: string
          description: >
            Id de la tarea. En `previous-visits` se usa el `workOrder` mientras
            no haya tareas cargadas: es lo que el técnico ve.
          example: TASK/294328/2026
        workOrder:
          type: string
          nullable: true
          description: Orden a la que pertenece. Se usa para pedir sus notas.
          example: ORDER/424900/2026
        occurredAt:
          type: string
          format: date-time
        reason:
          type: string
          description: Motivo. `""` si la operadora no lo trajo.
        closingNotes:
          type: string
          description: >
            Observación de cierre: concatenación de las notas de la tarea
            separadas por " · ". `""` mientras `notesLoaded` sea false.
        technician:
          type: string
          nullable: true
          description: >
            ⚠ **Siempre `null` con FSM**: `/workorder/tasks` no devuelve el
            técnico que cerró la tarea. Queda como punto abierto con la
            operadora; la UI muestra "—".
        result:
          type: string
          enum: [SATISFACTORIA, INSATISFACTORIA, PENDIENTE]
          description: >
            Valor **inferido** por el backend, no un dato de la operadora:
            `PENDIENTE` si la tarea no terminó, `INSATISFACTORIA` si el `status`
            o las notas contienen una palabra de `FSM_UNSATISFACTORY_KEYWORDS`,
            `SATISFACTORIA` en cualquier otro caso. Heurístico provisional a
            recalibrar con datos reales.
        notesLoaded:
          type: boolean
          description: >
            false cuando las notas todavía no se cargaron (respuesta de
            `previous-visits`). El cliente lo usa para saber que hay que llamar
            a `GET /workorders/tasks?workOrder=`.
    PreviousVisits:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClosedTask'
        totalOrders:
          type: integer
        brand:
          type: string
          enum: [telenews, seteinfo]
        degraded:
          $ref: '#/components/schemas/Degraded'
    UnsatisfactoryTasks:
      type: object
      properties:
        items:
          type: array
          description: Solo tareas cuyo `result` es INSATISFACTORIA.
          items:
            $ref: '#/components/schemas/ClosedTask'
        scanned:
          type: integer
          description: Órdenes finalizadas cuyas notas se leyeron.
        totalOrders:
          type: integer
          description: Órdenes totales de la cuenta.
        truncated:
          type: boolean
          description: true cuando `totalOrders > scanned`.
        brand:
          type: string
          enum: [telenews, seteinfo]
        degraded:
          $ref: '#/components/schemas/Degraded'
    AccountOrders:
      type: object
      properties:
        accountNumber:
          type: string
        brand:
          type: string
          enum: [telenews, seteinfo]
        client:
          type: object
          nullable: true
          description: Datos del cliente según la orden más reciente. `null` si no hay órdenes.
          properties:
            names:
              type: string
              nullable: true
            phoneNumber:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
            address:
              type: string
              nullable: true
            latitude:
              type: number
              format: double
              nullable: true
            longitude:
              type: number
              format: double
              nullable: true
        orders:
          type: array
          description: Ordenadas por `createdAt` descendente. Puede ser `[]`.
          items:
            type: object
            properties:
              workOrder:
                type: string
                example: ORDER/424900/2026
              task:
                type: string
                nullable: true
              state:
                type: string
                nullable: true
              externalProcess:
                type: string
                nullable: true
              cpartyId:
                type: string
                nullable: true
              createdAt:
                type: string
                format: date-time
                nullable: true
              endedAt:
                type: string
                format: date-time
                nullable: true
              finished:
                type: boolean
                description: Equivale a `endedAt != null` (`finish_date is not null`).
              note:
                type: string
                nullable: true
              latitude:
                type: number
                format: double
                nullable: true
              longitude:
                type: number
                format: double
                nullable: true
              address:
                type: string
                nullable: true
    WorkOrderTasks:
      type: object
      properties:
        workOrder:
          type: string
          example: ORDER/424900/2026
        brand:
          type: string
          enum: [telenews, seteinfo]
        tasks:
          type: array
          items:
            type: object
            properties:
              taskId:
                type: string
                example: TASK/294328/2026
              status:
                type: string
                nullable: true
                description: Literal de la operadora, sin traducir.
              businessKey:
                type: string
                nullable: true
              createdAt:
                type: string
                format: date-time
                nullable: true
              finishedAt:
                type: string
                format: date-time
                nullable: true
              result:
                type: string
                enum: [SATISFACTORIA, INSATISFACTORIA, PENDIENTE]
                description: Valor inferido por el backend (ver `ClosedTask.result`).
              notes:
                type: array
                description: Ordenadas por `createdAt` ascendente. Puede ser `[]`.
                items:
                  type: object
                  properties:
                    createdAt:
                      type: string
                      format: date-time
                      nullable: true
                    content:
                      type: string